refactor: move service defaults from shared/ into Host
The repository-root shared/ directory held one source file, linked into
src/Host with a Compile item reaching outside the project directory. That
arrangement is the Aspire template's shape for solutions with several
executable services; MailMcp has one, so it bought nothing and cost CI
coverage, because both workflows filter on src/** and tests/** and a file
outside those paths could change a production assembly without triggering
the build, unit-test, coverage, or formatting gates.

The file becomes src/Host/ServiceDefaultsExtensions.cs in the MailMcp.Host
namespace. Extensions in Microsoft.Extensions.Hosting was defensible for a
template scaffold in a shared folder; as a Host-owned file it has to follow
the repository rules on namespace/folder alignment and descriptive naming.
The AddServiceDefaults, ConfigureOpenTelemetry, AddDefaultHealthChecks, and
MapDefaultEndpoints member names are unchanged, so host startup registers
exactly the same defaults and /health and /alive stay development-only.

Record the rule in docs/architecture/solution-structure.md so the template's
shared/ scaffold is not reintroduced, and correct the specs and the Aspire
template's LICENSES.md row that named the old path.
