+++ LICENSES.md
@@ -30,7 +30,7 @@ Prohibited without explicit owner approval are strong copyleft or source-availab
 | Pgvector.EntityFrameworkCore 0.3.0 | EF Core vector mapping | MIT in NuGet package metadata and upstream repository metadata | Allowed. Permissive license supports both target distribution models with notice preservation. | <https://github.com/pgvector/pgvector-dotnet>, <https://www.nuget.org/packages/Pgvector.EntityFrameworkCore/> |
 | Microsoft.Agents.AI 1.15.0 | Agent and RAG orchestration, including ChatClientAgent and TextSearchProvider integration | MIT in NuGet package metadata | Allowed. Preview package terms and model/service terms remain separately reviewable on every update. | <https://github.com/microsoft/agent-framework>, <https://devblogs.microsoft.com/agent-framework/microsoft-agent-framework-version-1-0/>, <https://www.nuget.org/packages/Microsoft.Agents.AI/> |
 | Microsoft.Extensions.AI 10.8.1 and Microsoft.Extensions.AI.Abstractions 10.8.1 | Provider-neutral chat and embedding abstractions | MIT in NuGet package metadata | Allowed. Provider-specific packages remain separately reviewable dependencies. | <https://www.nuget.org/packages/Microsoft.Extensions.AI/>, <https://www.nuget.org/packages/Microsoft.Extensions.AI.Abstractions/>, <https://learn.microsoft.com/en-us/dotnet/ai/microsoft-extensions-ai> |
-| Aspire.Hosting.AppHost 13.4.6, Aspire.Hosting.PostgreSQL 13.4.6, and Aspire.ProjectTemplates 13.4.6 service defaults template | Development-time local orchestration for MailMcp and PostgreSQL, plus the shared service-defaults source scaffold for health checks, telemetry, resilience, and service discovery | MIT in NuGet package metadata and upstream repository | Allowed as development tooling/orchestration packages and source scaffold. Preserve required notices if distributed. | <https://www.nuget.org/packages/Aspire.Hosting.AppHost/13.4.6>, <https://www.nuget.org/packages/Aspire.Hosting.PostgreSQL/13.4.6>, <https://www.nuget.org/packages/Aspire.ProjectTemplates/13.4.6>, <https://github.com/dotnet/aspire> |
+| Aspire.Hosting.AppHost 13.4.6, Aspire.Hosting.PostgreSQL 13.4.6, and Aspire.ProjectTemplates 13.4.6 service defaults template | Development-time local orchestration for MailMcp and PostgreSQL, plus the service-defaults source scaffold for health checks, telemetry, resilience, and service discovery, vendored as `src/Host/ServiceDefaultsExtensions.cs` | MIT in NuGet package metadata and upstream repository | Allowed as development tooling/orchestration packages and source scaffold. Preserve required notices if distributed. | <https://www.nuget.org/packages/Aspire.Hosting.AppHost/13.4.6>, <https://www.nuget.org/packages/Aspire.Hosting.PostgreSQL/13.4.6>, <https://www.nuget.org/packages/Aspire.ProjectTemplates/13.4.6>, <https://github.com/dotnet/aspire> |
 | Microsoft.Extensions.Http.Resilience 10.6.0 and Microsoft.Extensions.ServiceDiscovery 10.6.0 | Shared service defaults for HTTP resilience and service discovery in executable service projects | MIT in NuGet package metadata | Allowed. Development/runtime framework libraries with permissive licensing. | <https://www.nuget.org/packages/Microsoft.Extensions.Http.Resilience/10.6.0>, <https://www.nuget.org/packages/Microsoft.Extensions.ServiceDiscovery/10.6.0> |
 | OpenTelemetry.Exporter.OpenTelemetryProtocol 1.15.3, OpenTelemetry.Extensions.Hosting 1.15.3, OpenTelemetry.Instrumentation.AspNetCore 1.15.2, OpenTelemetry.Instrumentation.Http 1.15.1, and OpenTelemetry.Instrumentation.Runtime 1.15.1 | Shared service defaults for OTLP export, OpenTelemetry hosting, ASP.NET Core, HTTP client, and runtime instrumentation | Apache-2.0 in NuGet package metadata | Allowed. Preserve required license and notice files, especially for exporters and instrumentation packages. | <https://www.nuget.org/packages/OpenTelemetry.Exporter.OpenTelemetryProtocol/1.15.3>, <https://www.nuget.org/packages/OpenTelemetry.Extensions.Hosting/1.15.3>, <https://www.nuget.org/packages/OpenTelemetry.Instrumentation.AspNetCore/1.15.2>, <https://www.nuget.org/packages/OpenTelemetry.Instrumentation.Http/1.15.1>, <https://www.nuget.org/packages/OpenTelemetry.Instrumentation.Runtime/1.15.1> |
 | OpenTelemetry .NET SDK and contrib packages | Traces, metrics, instrumentation, and exporters | Apache-2.0 | Allowed. Preserve required license and notice files, especially for exporters and contrib packages. | <https://github.com/open-telemetry/opentelemetry-dotnet>, <https://github.com/open-telemetry/opentelemetry-dotnet-contrib> |
