+++ src/Host/Host.csproj
@@ -17,7 +17,4 @@
     <ProjectReference Include="..\Infrastructure\Infrastructure.csproj" />
     <ProjectReference Include="..\Mcp\Mcp.csproj" />
   </ItemGroup>
-  <ItemGroup>
-    <Compile Include="..\..\shared\Extensions.cs" Link="shared\Extensions.cs" />
-  </ItemGroup>
 </Project>
